+++ src/app/private/pages/editarIdoso.tsx
@@ -27,6 +27,7 @@ import database from "../../db";
 import { Collection, Q } from "@nozbe/watermelondb";
 import Idoso from "../../model/Idoso";
 import { getTipoSanguineoOptions } from "../../shared/helpers/useNotification";
+import styles from "../../components/style/styles";
 
 
 interface IErrors {
@@ -344,7 +345,7 @@ export default function EditarIdoso() {
          {showLoadingApagar ? (
            <ActivityIndicator size="small" color="#FF7F7F" />
          ) : (
-           <Text style={styles.apagar}>Apagar Idoso</Text>
+           <Text style={styles.apagarEditar}>Apagar Idoso</Text>
          )}
        </Pressable>
 
@@ -361,99 +362,3 @@ export default function EditarIdoso() {
  );
 }
 
-
-const styles = StyleSheet.create({
- voltar: {
-   marginTop: 5,
- },
- formControl: {
-   flexDirection: "column",
-   width: 320,
-   alignItems: "flex-start",
-   alignSelf: "center",
-   marginTop: 10,
- },
- formControl2: {
-   flexDirection: "row",
-   width: 320,
-   alignItems: "flex-start",
-   alignSelf: "center",
-   marginTop: 10,
- },
- button: {
-   width: "80%",
-   maxWidth: 350,
-   paddingVertical: 16,
-   paddingHorizontal: 26,
-   alignItems: "center",
-   borderRadius: 20,
-   backgroundColor: "#2CCDB5",
-   textAlign: "center",
- },
- field: {
-   flexDirection: "row",
-   width: 320,
-   borderBottomWidth: 1,
-   borderBottomColor: "#AFB1B6",
-   paddingBottom: 5,
-   alignSelf: "center",
-   marginBottom: 5,
- },
- fieldBlood: {
-   flexDirection: "row",
-   width: 320,
-   borderBottomWidth: 1,
-   borderBottomColor: "#AFB1B6",
-   paddingBottom: 5,
-   alignSelf: "center",
-   marginBottom: 5,
- },
- iconInput: {
-   width: "10%",
-   alignSelf: "center",
-   marginLeft: 10,
- },
- iconInput2: {
-   width: "10%",
-   marginTop: "7%",
-   marginLeft: 10,
- },
- bloodInput: {
-   paddingLeft: 10,
-   color: "#05375a",
-   width: "80%",
-   fontSize: 17,
- },
- bloodIcon: {
-   width: "10%",
- },
- textInput: {
-   width: "90%",
-   paddingLeft: 10,
-   color: "#05375a",
-   fontSize: 17,
- },
- arrow: {
-   alignSelf: "flex-start",
- },
- linkButton: {
-   marginTop: 60,
-   marginBottom: 40,
-   alignItems: "center",
- },
- apagar: {
-   color: "#FF7F7F",
-   alignSelf: "center",
-   fontSize: 18,
-   fontWeight: "600",
-   marginBottom: 60,
-   alignItems: "center",
- },
- dropdown: {
-   width: 300,
-   borderWidth: 0,
-   paddingLeft: 10,
-   color: "#05375A",
-   fontSize: 17,
- },
-});
